+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0-RC2] - 2025-10-17
+
+### Added
+- Improve error handling for localized data import
+- Add Pester tests and GitHub workflow
+
+### Changed
+- Update GitHub Release action for new folder structures
+
 ## [1.5.0-RC1] - 2025-10-10
 
 ### Added